create client.py + test fix 5
